migrate to new phase
